Update my code 2

+++ basic_math.py
@@ -42,7 +42,6 @@ def get_median(number_list):
     if length % 2:
         median = number_list[length / 2]
     else:
-        median = float(
-            (number_list[length / 2 - 1] + number_list[length / 2])) / 2
+        median = number_list[length / 2 - 1] + number_list[length / 2] / 2
 
     return median
